refactor(p4-3/3): shared prospect constants/parsers → prospectShared.js
